[ADD] better AbstractCSVLogger - Add new BlockchainReorgLogger that can work with any other PoW blockchain without the need for any special events - Improve PoW networks and nodes model - Add new factory methods to BitcoinBlock - Add genesis block to all consensus algorithms configs - Add new logger for uncle block counting - Add new class for PoW single parent blocks

+++ src/main/java/jabs/network/networks/bitcoin/BitcoinGlobalProofOfWorkNetwork.java
@@ -36,7 +36,7 @@ public class BitcoinGlobalProofOfWorkNetwork<R extends Enum<R>> extends
     }
 
     @Override
-    public BitcoinMinerNode createSampleMiner(Simulator simulator, int nodeID,  long hashPower,
+    public BitcoinMinerNode createSampleMiner(Simulator simulator, int nodeID, double hashPower,
                                               BitcoinBlockWithoutTx genesisBlock,
                                               ChainBasedConsensusConfig chainBasedConsensusConfig) {
         R region = (R) this.sampleRegion();
