+++ src/main/java/jabs/network/node/nodes/ethereum/EthereumCasperMiner.java
@@ -9,7 +9,7 @@ import jabs.simulator.Simulator;
 
 public class EthereumCasperMiner extends EthereumMinerNode {
     public EthereumCasperMiner(Simulator simulator, Network network, int nodeID, long downloadBandwidth,
-                               long uploadBandwidth, long hashPower, EthereumBlock genesisBlock,
+                               long uploadBandwidth, double hashPower, EthereumBlock genesisBlock,
                                CasperFFGConfig casperFFGConfig) {
         super(simulator, network, nodeID, downloadBandwidth, uploadBandwidth, hashPower,
                 new CasperFFG<>(new LocalBlockTree<>(genesisBlock), casperFFGConfig));
